requirements.txt
Flask==3.1.3
requests==2.32.5
python-dotenv==1.2.2
mysql-connector-python
pytest==9.0.2
